FEATURE: Add rudimentary pointer capture functionality
This is useful for when you go out of bounds of a hit test and wish to still receive pointer events. For example, often when scrolling with a scroll thumb, you will bring the mouse to the middle of the screen rather than staying hovered over the scrollbar, so that still needs to receive pointer events despite not being in bounds.